backend/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0